requirements.txt
# Core dependencies
asyncio
pandas>=2.0.0
numpy>=1.24.0
TA-Lib>=0.4.28
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Exchange API
pybit>=5.6.0

# Optional: API monitoring
flask>=3.0.0
flask-cors>=4.0.0

# Logging and utilities
colorlog>=6.7.0

# Data storage (optional)
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Technical analysis (additional)
scipy>=1.11.0
scikit-learn>=1.3.0

# Websocket support (for real-time data)
websockets>=11.0
websocket-client>=1.6.0

# Performance monitoring
psutil>=5.9.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0